Move types under the `type` package
