+++ base/src/main/java/io/spine/code/proto/FieldDeclaration.java
@@ -31,7 +31,9 @@ import io.spine.code.java.ClassName;
 import io.spine.logging.Logging;
 import io.spine.option.EntityOption;
 import io.spine.option.OptionsProto;
+import io.spine.type.EnumType;
 import io.spine.type.KnownTypes;
+import io.spine.type.MessageType;
 import io.spine.type.TypeName;
 import io.spine.type.TypeUrl;
 import io.spine.type.UnknownTypeException;
