+++ base/src/main/java/io/spine/code/proto/LocationPath.java
@@ -24,10 +24,14 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.protobuf.DescriptorProtos.FileDescriptorProto;
 import com.google.protobuf.DescriptorProtos.SourceCodeInfo;
 import com.google.protobuf.DescriptorProtos.SourceCodeInfo.Location;
+import com.google.protobuf.Descriptors.Descriptor;
+import io.spine.type.MessageType;
 
+import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Deque;
 import java.util.List;
 
 import static com.google.common.base.Preconditions.checkArgument;
@@ -40,7 +44,7 @@ import static java.lang.String.format;
  * <p>A location path represents {@linkplain Location#getPathList() list of
  * integers}, that used to identify a {@linkplain Location location} in a ".proto" file.
  */
-final class LocationPath {
+public final class LocationPath {
 
     private final List<Integer> path;
 
@@ -66,6 +70,25 @@ final class LocationPath {
     @SuppressWarnings("unused") // Included for referencing `Location` in Javadoc.
     LocationPath(Location location) {
         this(location.getPathList());
+    }
+
+    /**
+     * Creates an instance for the passed message descriptor.
+     */
+    public static LocationPath fromMessage(Descriptor descriptor) {
+        LocationPath path = new LocationPath();
+        path.add(FileDescriptorProto.MESSAGE_TYPE_FIELD_NUMBER);
+        if (!MessageType.isTopLevel(descriptor)) {
+            Deque<Integer> parentPath = new ArrayDeque<>();
+            Descriptor containingType = descriptor.getContainingType();
+            while (containingType != null) {
+                parentPath.addFirst(containingType.getIndex());
+                containingType = containingType.getContainingType();
+            }
+            path.addAll(parentPath);
+        }
+        path.add(descriptor.getIndex());
+        return path;
     }
 
     /**
@@ -106,9 +129,9 @@ final class LocationPath {
     }
 
     /**
-     * Converts the instance to the {@link SourceCodeInfo.Location} in the given file.
+     * Converts the instance to the {@code SourceCodeInfo.Location} instance in the given file.
      */
-    SourceCodeInfo.Location toLocation(FileDescriptorProto file) {
+    public SourceCodeInfo.Location toLocation(FileDescriptorProto file) {
         List<Integer> thisPath = toList();
         for (SourceCodeInfo.Location location : file.getSourceCodeInfo()
                                                     .getLocationList()) {
