+++ base/src/main/java/io/spine/type/TypeName.java
@@ -25,7 +25,6 @@ import com.google.errorprone.annotations.Immutable;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.GenericDescriptor;
 import com.google.protobuf.Message;
-import io.spine.code.proto.Type;
 import io.spine.value.StringTypeValue;
 
 import java.util.List;
