+++ base/src/test/java/io/spine/code/proto/enrichment/EnrichmentTypeTest.java
@@ -25,7 +25,6 @@ import com.google.common.truth.IterableSubject;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.Message;
-import io.spine.code.proto.MessageType;
 import io.spine.test.code.enrichment.type.EttAlternativeFieldNames;
 import io.spine.test.code.enrichment.type.EttFieldSelection;
 import io.spine.test.code.enrichment.type.EttOnAnotherPackageMessage;
@@ -39,6 +38,7 @@ import io.spine.test.code.enrichment.type.user.permission.EttPermissionGrantedEv
 import io.spine.test.code.enrichment.type.user.permission.EttPermissionRevokedEvent;
 import io.spine.test.code.enrichment.type.user.sharing.EttSharingRequestApproved;
 import io.spine.test.code.enrichment.type.user.sharing.EttSharingRequestSent;
+import io.spine.type.MessageType;
 import io.spine.type.TypeName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
