+++ tools/errorprone-checks/src/main/java/io/spine/tools/check/vbuilder/fixer/FixGenerator.java
@@ -32,7 +32,7 @@ import com.sun.tools.javac.code.Type;
 import com.sun.tools.javac.tree.JCTree.JCExpression;
 import com.sun.tools.javac.tree.JCTree.JCFieldAccess;
 import com.sun.tools.javac.tree.JCTree.JCIdent;
-import io.spine.code.proto.MessageType;
+import io.spine.type.MessageType;
 
 import static com.google.errorprone.fixes.SuggestedFixes.prettyType;
 import static com.google.errorprone.util.ASTHelpers.enclosingClass;
