+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/annotation/MethodNameAnnotator.java
@@ -24,8 +24,8 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.Descriptors.FileDescriptor;
 import io.spine.code.java.ClassName;
-import io.spine.code.proto.Type;
 import io.spine.code.proto.TypeSet;
+import io.spine.type.Type;
 import org.jboss.forge.roaster.model.Method;
 import org.jboss.forge.roaster.model.impl.AbstractJavaSource;
 import org.jboss.forge.roaster.model.source.JavaClassSource;
