+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/validation/VBuilderMethods.java
@@ -28,9 +28,9 @@ import com.squareup.javapoet.CodeBlock;
 import com.squareup.javapoet.MethodSpec;
 import io.spine.code.java.SimpleClassName;
 import io.spine.code.proto.FieldDeclaration;
-import io.spine.code.proto.MessageType;
 import io.spine.protobuf.Messages;
 import io.spine.tools.compiler.field.type.FieldType;
+import io.spine.type.MessageType;
 
 import javax.lang.model.element.Modifier;
 import java.util.Collection;
