+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/validation/ValidationRulesWriter.java
@@ -23,8 +23,8 @@ package io.spine.tools.compiler.validation;
 import com.google.protobuf.DescriptorProtos.DescriptorProto;
 import io.spine.code.properties.PropertiesWriter;
 import io.spine.code.proto.FileSet;
-import io.spine.code.proto.MessageType;
 import io.spine.logging.Logging;
+import io.spine.type.MessageType;
 import io.spine.type.TypeName;
 import io.spine.validate.rule.ValidationRules;
 import org.checkerframework.checker.nullness.qual.Nullable;
