+++ tools/proto-js-plugin/src/main/java/io/spine/js/generate/index/KnownTypesMap.java
@@ -23,11 +23,11 @@ package io.spine.js.generate.index;
 import com.google.common.collect.Maps;
 import io.spine.code.js.TypeName;
 import io.spine.code.proto.FileSet;
-import io.spine.code.proto.Type;
 import io.spine.code.proto.TypeSet;
 import io.spine.js.generate.Snippet;
 import io.spine.js.generate.output.CodeLines;
 import io.spine.js.generate.output.snippet.MapExportSnippet;
+import io.spine.type.Type;
 import io.spine.type.TypeUrl;
 
 import java.util.List;
