+++ tools/proto-js-plugin/src/test/java/io/spine/js/generate/AppendTypeUrlGetterTest.java
@@ -24,15 +24,15 @@ import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.EnumDescriptor;
 import com.google.protobuf.Descriptors.FileDescriptor;
 import com.google.protobuf.Descriptors.ServiceDescriptor;
-import io.spine.code.proto.MessageType;
-import io.spine.code.proto.ServiceType;
-import io.spine.code.proto.Type;
 import io.spine.js.generate.output.CodeLines;
 import io.spine.js.generate.output.snippet.Method;
 import io.spine.js.generate.typeurl.OuterMessage;
 import io.spine.js.generate.typeurl.OuterMessage.NestedEnum;
 import io.spine.js.generate.typeurl.OuterMessage.NestedMessage;
 import io.spine.js.generate.typeurl.TopLevelEnum;
+import io.spine.type.MessageType;
+import io.spine.type.ServiceType;
+import io.spine.type.Type;
 import io.spine.type.TypeUrl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
