+++ tools/proto-js-plugin/src/test/java/io/spine/js/generate/parse/GenerateKnownTypeParsersTest.java
@@ -29,7 +29,6 @@ import io.spine.code.js.FileName;
 import io.spine.code.js.TypeName;
 import io.spine.code.proto.FileDescriptors;
 import io.spine.code.proto.FileSet;
-import io.spine.code.proto.MessageType;
 import io.spine.code.proto.TypeSet;
 import io.spine.js.generate.TaskId;
 import io.spine.js.generate.given.GivenProject;
@@ -37,6 +36,7 @@ import io.spine.js.generate.output.CodeLines;
 import io.spine.js.generate.output.snippet.Comment;
 import io.spine.js.generate.output.snippet.Import;
 import io.spine.option.OptionsProto;
+import io.spine.type.MessageType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
