+++ tools/protoc-plugin/src/main/java/io/spine/tools/protoc/insert/IdentityParameter.java
@@ -21,7 +21,7 @@
 package io.spine.tools.protoc.insert;
 
 import com.google.errorprone.annotations.Immutable;
-import io.spine.code.proto.Type;
+import io.spine.type.Type;
 
 /**
  * The message interface parameter whose value is the target {@code Message} itself.
