+++ tools/protoc-plugin/src/main/java/io/spine/tools/protoc/insert/MessageAndInterface.java
@@ -26,9 +26,9 @@ import com.google.protobuf.DescriptorProtos.DescriptorProto;
 import com.google.protobuf.DescriptorProtos.FileDescriptorProto;
 import com.google.protobuf.Descriptors;
 import com.google.protobuf.compiler.PluginProtos.CodeGeneratorResponse.File;
-import io.spine.code.proto.MessageType;
 import io.spine.option.IsOption;
 import io.spine.tools.protoc.CompilerOutput;
+import io.spine.type.MessageType;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.util.Collection;
